Migrate to AGP 9 KMP android plugin; replace bignum with Double (#219)
* Migrate to AGP 9 KMP android plugin and replace bignum with Double

Switch all modules from the deprecated `androidLibrary {}` block to the
new `android {}` block required by AGP 9's `com.android.kotlin.multiplatform.library`
plugin. Restore `withAndroidTarget()` alongside the existing `withCompilations`
workaround for issue 442950553 so the IDE model is set up correctly.

The new plugin requires KMP libraries to publish an androidTarget artifact.
`com.ionspin.kotlin:bignum` only publishes jvm/native variants, so replace
its `BigDecimal` usage in the scripting module with stdlib `Double`. Precision
is more than adequate for WSL scripting arithmetic (counters, timers, comparisons).


* Add WslValue.toText() for script-facing string conversion (#220)

Introduce an explicit `toText()` on WslValue (parallel to toNumber/toBoolean)
as the way scripts read a value as text - variable substitution, concatenation,
comparison, contains, indexing keys, etc. `toString()` now delegates to it and
is reserved for debugging/logging. This removes the reliance on `toString()`
for script semantics.

Fixes two latent WslVariable bugs surfaced by the change:
- toText()/toString() returned the identity string (WslVariable@hash), so
  built-in variables like %right and %character substituted to garbage.
- toNumber() called this.toString() instead of the wrapped value, so a
  string-backed variable always converted to 0.0.

Also tightens two nullable lookupVariable(...).toString() sites that silently
produced "null" to the null-safe ?.toText() ?: "" pattern.

Adds end-to-end example script tests (WslScriptTest) that parse and run whole
scripts and assert on computed values, plus shared parseScript/runToCompletion
test helpers and WslVariable regression cases.


---------

+++ scripting/src/commonMain/kotlin/warlockfe/warlock3/scripting/wsl/WslTimer.kt
@@ -1,14 +1,12 @@
 package warlockfe.warlock3.scripting.wsl
 
-import com.ionspin.kotlin.bignum.decimal.BigDecimal
-import com.ionspin.kotlin.bignum.decimal.toBigDecimal
 import kotlin.time.Clock
 
 class WslTimer : WslNumeric() {
     private val startTime = Clock.System.now()
 
-    override fun toNumber(): BigDecimal {
+    override fun toNumber(): Double {
         val duration = Clock.System.now() - startTime
-        return duration.inWholeSeconds.toBigDecimal()
+        return duration.inWholeSeconds.toDouble()
     }
 }
